Module janus_core::time

Utilities for timestamps and durations.

Structs

MockClock

A mock clock for use in testing. Clones are identical: all clones of a given MockClock will be controlled by a controller retrieved from any of the clones.

RealClock

A real clock returns the current time relative to the Unix epoch.

Traits

Clock

A clock knows what time it currently is.

DurationExt

Extension methods on Duration.

IntervalExt

Extension methods on Interval.

TimeExt

Extension methods on Time.