Trait janus_core::time::DurationExt

pub trait DurationExt: Sized {
    // Required methods
    fn as_chrono_duration(&self) -> Result<Duration, Error>;
    fn add(&self, duration: &Duration) -> Result<Self, Error>;
    fn from_microseconds(microseconds: u64) -> Self;
    fn as_microseconds(&self) -> Result<u64, Error>;
    fn from_minutes(minutes: u64) -> Result<Self, Error>;
    fn from_hours(hours: u64) -> Result<Self, Error>;
    fn round_up(&self, time_precision: &Duration) -> Result<Self, Error>;
}

Extension methods on Duration.

Required Methods

fn as_chrono_duration(&self) -> Result<Duration, Error>

Convert this Duration into a chrono::Duration.

fn add(&self, duration: &Duration) -> Result<Self, Error>

Add this duration with another duration.

fn from_microseconds(microseconds: u64) -> Self

Create a duration from a number of microseconds. The time will be rounded down to the next second.

fn as_microseconds(&self) -> Result<u64, Error>

Get the number of microseconds this duration represents. Note that the precision of this type is one second, so this method will always return a multiple of 1,000,000 microseconds.

fn from_minutes(minutes: u64) -> Result<Self, Error>

Create a duration representing the provided number of minutes.

fn from_hours(hours: u64) -> Result<Self, Error>

Create a duration representing the provided number of hours.

fn round_up(&self, time_precision: &Duration) -> Result<Self, Error>

Return a duration representing this duration rounded up to the next largest multiple of time_precision, or the same duration if it already a multiple.