Trait janus_core::time::TimeExt

pub trait TimeExt: Sized {
    // Required methods
    fn to_batch_interval_start(
        &self,
        time_precision: &Duration
    ) -> Result<Self, Error>;
    fn as_naive_date_time(&self) -> Result<NaiveDateTime, Error>;
    fn from_naive_date_time(time: &NaiveDateTime) -> Self;
    fn add(&self, duration: &Duration) -> Result<Self, Error>;
    fn sub(&self, duration: &Duration) -> Result<Self, Error>;
    fn difference(&self, other: &Self) -> Result<Duration, Error>;
    fn is_after(&self, time: &Time) -> bool;
}

Extension methods on Time.

Required Methods

fn to_batch_interval_start( &self, time_precision: &Duration ) -> Result<Self, Error>

Compute the start of the batch interval containing this Time, given the task time precision.

fn as_naive_date_time(&self) -> Result<NaiveDateTime, Error>

Convert this Time into a NaiveDateTime, representing an instant in the UTC timezone.

fn from_naive_date_time(time: &NaiveDateTime) -> Self

Convert a NaiveDateTime representing an instant in the UTC timezone into a Time.

fn add(&self, duration: &Duration) -> Result<Self, Error>

Add the provided duration to this time.

fn sub(&self, duration: &Duration) -> Result<Self, Error>

Subtract the provided duration from this time.

fn difference(&self, other: &Self) -> Result<Duration, Error>

Get the difference between the provided other and self. self must be after other.

fn is_after(&self, time: &Time) -> bool

Returns true if this Time occurs after time.