Module janus_core::vdaf

Modules

vdaf_dp_strategies

Contains dedicated enums which describe the differential privacy strategies of a given VDAF. If a VDAF only supports a single strategy, such as for example NoDifferentialPrivacy, then no enum is required.

Enums

Prio3FixedPointBoundedL2VecSumBitSize

Bitsize parameter for the Prio3FixedPointBoundedL2VecSum VDAF.

VdafInstance

Identifiers for supported VDAFs, corresponding to definitions in draft-irtf-cfrg-vdaf-03 and implementations in prio::vdaf::prio3.

Constants

VERIFY_KEY_LENGTH

The length of the verify key parameter for Prio3 & Poplar1 VDAF instantiations.