Function janus_core::retries::retry_http_request

pub async fn retry_http_request<RequestFn, ResultFuture>(
    backoff: ExponentialBackoff,
    request_fn: RequestFn
) -> Result<Response, Result<Response, Error>>where
    RequestFn: Fn() -> ResultFuture,
    ResultFuture: Future<Output = Result<Response, Error>>,

Executes the provided request function and awaits the returned future, retrying using the parameters in the provided ExponentialBackoff if the reqwest::Error returned by request_fn is:

a timeout

a problem establishing a connection

an HTTP status code indicating a server error

If the request eventually succeeds, the value returned by request_fn is returned. If an unretryable failure occurs or enough transient failures occur, then Err(ret) is returned, where ret is the Result<reqwest::Response, reqwest::Error> returned by the last call to request_fn. Retryable failures are logged.

TODOs:

This function could take a list of HTTP status codes that should be considered retryable, so that a caller could opt to retry when it sees 408 Request Timeout or 429 Too Many Requests, but since none of the servers this is currently used to communicate with ever return those statuses, we don’t yet need that feature.